requirements.txt
langchain
langchain-community
langgraph
sentence-transformers
diskcache
pymupdf
fastapi
uvicorn
python-dotenv
ipykernel
pandas
faiss-cpu
rank-bm25
arxiv
